Renames ExitServlet to ExitJvmServlet
