+++ clusterbench-ee5-web/src/main/java/org/jboss/test/clusterbench/web/exitjvm/ExitJvmServlet.java
@@ -1,4 +1,4 @@
-package org.jboss.test.clusterbench.web.exit;
+package org.jboss.test.clusterbench.web.exitjvm;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
@@ -6,7 +6,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class ExitServlet extends HttpServlet {
+public class ExitJvmServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
