+++ clusterbench-ee6-web/src/main/webapp/index.html
@@ -39,7 +39,7 @@
             <li><strong>AverageSystemLoadServlet</strong> <a href="averagesystemload?milliseconds=20000&threads=4">/averagesystemload?milliseconds=20000&threads=4</a></li>
             <li><strong>SendTrafficLoadServlet</strong> <a href="sendtrafficload?kilobytes=100">/sendtrafficload?kilobytes=100</a></li>
             <li><strong>BusyConnectorsLoadServlet</strong> <a href="busyconnectorsload?milliseconds=10000">/busyconnectorsload?milliseconds=10000</a></li>
-            <li><strong>ExitServlet</strong> <a href="exit">/exit</a></li>
+            <li><strong>ExitJvmServlet</strong> <a href="exitjvm">/exitjvm</a></li>
             <li><strong>ReceiveTrafficLoadServlet</strong> Set <i>data</i> in POST method to /receivetrafficload. See receiveTrafficLoadMetricManualTest in clusterbench-common.</li>
         </ul>
