+++ clusterbench-ee7-web/src/main/java/org/jboss/test/clusterbench/web/exitjvm/ExitJvmServlet.java
@@ -1,4 +1,4 @@
-package org.jboss.test.clusterbench.web.exit;
+package org.jboss.test.clusterbench.web.exitjvm;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,8 +7,8 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(name = "ExitServlet", urlPatterns = "/exit")
-public class ExitServlet extends HttpServlet {
+@WebServlet(name = "ExitJvmServlet", urlPatterns = "/exitjvm")
+public class ExitJvmServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         Runtime.getRuntime().halt(0);
